use falloff list evaluation in Offset Matrix node

+++ animation_nodes/algorithms/matrices/rotation.pxd
@@ -5,13 +5,13 @@ from ... math cimport (
 
 from ... data_structures cimport (
     Matrix4x4List,
-    DoubleList,
+    FloatList,
     VirtualEulerList
 )
 
 ctypedef void (*RotateFunction)(Matrix4 *target, Matrix4 *m, Euler3 *v)
 
 cpdef getRotatedMatrixList(Matrix4x4List matrices, str type,
-                           VirtualEulerList rotations, DoubleList influences)
+                           VirtualEulerList rotations, FloatList influences)
 
 cdef RotateFunction getRotateFunction(str type) except *
